추천 페이지 공통 컴포넌트 분리 작업
