+++ src/App.js
@@ -15,6 +15,8 @@ import ReSign from 'components/member/mypage/ReSign/ReSign';
 import BoardAddEditPage from 'pages/blog/BoardAddEditPage/BoardAddEditPage';
 import BoardDetail from 'pages/blog/BoardDetailPage/BoardDetail';
 import AllExercisePages from 'pages/recommend/exercise/AllExercisePages';
+import RecommendExercise from 'pages/recommend/exercise/RecommendExercise/RecommendExercise';
+
 
 function App() {
     return (
@@ -23,6 +25,7 @@ function App() {
                 <Header />
 
                 <div className={styles.layout}>
+                    <RecommendExercise/>
                     <Routes>
                         <Route path={"/"} element={<MainPage/>} />
                         <Route path={"/signUp"} element={<SignUpPage/>} />
