--- src/components/recommend/MessageForm/MessageForm.module.css
@@ -1,25 +1,0 @@
-/* .container {
-    width: 40rem;
-    height: 8rem;
-    border: solid 2px #d9d9d9;
-    border-radius: 5px;
-    box-shadow: 0px 0px 5px #d9d9d9;
-    margin-bottom : 3rem;
-    padding: 0.5rem;
-} */
-
-.container p {
-    text-align: center;
-    font-weight: 600;
-    /* font-size: 1rem; */
-    /* padding: 1rem; */
-}
-
-.burnKcal {
-    color: #081f5c;
-}
-
-.kcal {
-    color: red;
-    margin: 0 0.3rem;
-}
