+++ src/components/recommend/RepeatDay/RepeatDay.module.css
@@ -1,5 +1,5 @@
 .form {
-    margin : 10rem auto;
+    margin : 5rem auto;
     width: 45rem;
     padding: 2rem;
 }
