+++ src/components/recommend/RoutineRecommendForm/RoutineRecommendForm.module.css
@@ -1,0 +1,37 @@
+.container {
+    margin: 10rem auto;
+}
+
+.container button {
+    float: right;
+}
+
+.title {
+    font-size : 2rem;
+    font-weight : bold;
+    color : #334eac;
+    margin : 5rem;
+    text-align: center;
+}
+
+.category {
+    display: flex;
+    justify-content: center;
+    margin: 3rem auto;
+}
+
+.registButton {
+    display: block;
+    margin-top: 3rem;
+}
+
+.message {
+    text-align: center;
+    font-weight : bold;
+    margin : 1rem 0;
+}
+
+.burnKcal{
+    color : #081F5C;
+    margin : 0.5rem 0;
+}
