+++ src/pages/recommend/exercise/ExerciseInputInfo/ExerciseInputInfo.jsx
@@ -1,10 +1,10 @@
-import RecommendForm from 'components/recommend/RecommendForm/RecommendForm';
+import RoutineInfoForm from 'components/recommend/RoutineInfoForm/RoutineInfoForm';
 import styles from './ExerciseInputInfo.module.css';
 
 function ExerciseInputInfo({ goToNext, formData, setFormData }) {
     return (
         <div className={styles.container}>
-            <RecommendForm
+            <RoutineInfoForm
                 title="운동 추천"
                 goToNext={goToNext}
                 formData={formData}
