+++ src/pages/recommend/exercise/ExerciseRepeatDay/ExerciseRepeatDay.module.css
@@ -1,0 +1,5 @@
+.container {
+    display : flex;
+    justify-content: center;
+    margin : auto;
+}
